nexus-reussite/backend/requirements-dev.txt
# ==========================================
# NEXUS RÉUSSITE - DEVELOPMENT REQUIREMENTS  
# ==========================================
# Dépendances supplémentaires pour le développement
# À utiliser avec: pip install -r requirements-production.txt -r requirements-dev.txt

# === TESTS ===
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-xdist==3.5.0         # Tests parallèles
pytest-mock==3.12.0         # Mock facilité
factory-boy==3.3.0          # Factories pour les tests
faker==20.1.0               # Données de test

# === QUALITÉ DE CODE ===
black==25.1.0               # Formatage automatique
flake8==6.1.0               # Linting
pylint==3.0.3               # Analyse statique avancée
mypy==1.7.1                 # Type checking
isort==5.13.2               # Tri des imports
pre-commit==3.6.0           # Git hooks

# === SÉCURITÉ DÉVELOPPEMENT ===
bandit==1.8.6               # Audit de sécurité
safety==2.3.5               # Audit des vulnérabilités
pip-audit==2.6.1            # Audit des dépendances

# === DOCUMENTATION ===
sphinx==7.2.6               # Génération documentation
sphinx-rtd-theme==1.3.0     # Thème Read the Docs
sphinx-autodoc-typehints==1.25.2

# === OUTILS DE DÉVELOPPEMENT ===
ipython==8.18.1             # REPL amélioré
ipdb==0.13.13               # Débogueur amélioré
python-dotenv==1.0.0        # Variables d'environnement (déjà en prod)
watchdog==3.0.0             # Auto-reload fichiers

# === PROFILING & PERFORMANCE ===
memory-profiler==0.61.0     # Profiling mémoire
line-profiler==4.1.1        # Profiling ligne par ligne
py-spy==0.3.14              # Sampling profiler

# === UTILITAIRES ===
httpie==3.2.2               # Client HTTP CLI
rich==13.7.0                # Output coloré et formaté
typer==0.9.0                # CLI moderne
pydantic-extra-types==2.4.1 # Types Pydantic étendus

# === BASE DE DONNÉES DÉVELOPPEMENT ===
sqlite3                     # Inclus avec Python
fakeredis==2.20.1          # Redis en mémoire pour les tests

# === LINTING ADDITIONNEL ===
flake8-docstrings==1.7.0    # Vérification docstrings
flake8-import-order==0.18.2 # Ordre des imports
flake8-bugbear==23.12.2     # Détection de bugs
flake8-comprehensions==3.14.0

# === TYPE CHECKING EXTENSIONS ===
types-requests==2.31.0.20231231
types-redis==4.6.0.11
types-python-dateutil==2.8.19.14
